+++ mount_slash/fidc_cli.h
@@ -38,7 +38,7 @@ struct fidc_membh;
 
 struct fcmh_cli_info_file {
 	struct srt_inode	 inode;
-	uint64_t		 xattrsize;
+	uint32_t	 	 xattrsize;
 	int			 idxmap[SL_MAX_REPLICAS];
 	int			 mapstircnt;
 };
@@ -146,7 +146,7 @@ int	fcmh_checkcreds_ctx(struct fidc_membh *,
 
 int	msl_fcmh_fetch_inode(struct fidc_membh *);
 void	msl_fcmh_stash_inode(struct fidc_membh *, struct srt_inode *);
-void	msl_fcmh_stash_xattrsize(struct fidc_membh *, uint64_t);
+void	msl_fcmh_stash_xattrsize(struct fidc_membh *, uint32_t);
 
 #define fidc_lookup_load(fid, fcmhp, pfcc)				\
 	_fidc_lookup_load(PFL_CALLERINFOSS(SLSS_FCMH), (fid),		\
